separate useCustomerForm
